+++ CHANGES.rst
@@ -9,6 +9,13 @@ Version 1.1.dev1 (Unreleased)
 Enhancements
 ~~~~~~~~~~~~
 
+* Extended methods :meth:`Structure.view_fields`, :meth:`Sequence.view_fields`
+  and :meth:`Pointer.view_fields` with keyword `fieldnames` to support customized
+  fieldnames for the selected field *attributes*.
+* Extended method :meth:`Container.to_json` with keyword `fieldnames` to support
+  customized fieldnames for the selected field *attributes*.
+
+
 Backward-incompatible Changes
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~